Change inode number type from `usize` to `i32`.

+++ lib/ext4/src/inode/file.rs
@@ -22,7 +22,7 @@ unsafe impl Sync for ExtFileInode {}
 impl ExtFileInode {
     pub fn new(superblock: Arc<dyn SuperBlock>, file: ExtFile) -> Arc<Self> {
         let fsize = file.size();
-        let meta = InodeMeta::new(file.ino() as usize, superblock);
+        let meta = InodeMeta::new(file.ino() as i32, superblock);
         meta.inner.lock().size = fsize as usize;
         Arc::new(Self {
             meta,
